tool/msgpact_version/pyproject.toml
[project]
name = "msgpact_version"
version = "0.1.5"
description = "A CLI tool to update versions in msgpact sub projects."
requires-python = ">=3.11"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
version = "0.1.1"

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.0.0"
toml = "^0.10.2"
lxml = "^5.3.1"
pyyaml = "^6.0.2"

[tool.poetry.scripts]
msgpact-version = "msgpact_version.cli:main"
